+++ ajax_select/admin.py
@@ -1,7 +1,5 @@
-
-
-from ajax_select.fields import autoselect_fields_check_can_add
 from django.contrib import admin
+from ajax_select.fields import autoselect_fields_check_can_add
 
 
 class AjaxSelectAdmin(admin.ModelAdmin):
@@ -15,9 +13,17 @@ class AjaxSelectAdmin(admin.ModelAdmin):
         return form
 
 
-class AjaxSelectAdminTabularInline(admin.TabularInline):
+class AjaxSelectAdminInlineFormsetMixin(object):
 
     def get_formset(self, request, obj=None, **kwargs):
-        fs = super(AjaxSelectAdminTabularInline, self).get_formset(request, obj, **kwargs)
+        fs = super(AjaxSelectAdminInlineFormsetMixin, self).get_formset(request, obj, **kwargs)
         autoselect_fields_check_can_add(fs.form, self.model, request.user)
         return fs
+
+
+class AjaxSelectAdminTabularInline(AjaxSelectAdminInlineFormsetMixin, admin.TabularInline):
+    pass
+
+
+class AjaxSelectAdminStackedInline(AjaxSelectAdminInlineFormsetMixin, admin.StackedInline):
+    pass