+++ ajax_select/static/ajax_select/js/ajax_select.js
@@ -1,9 +1,9 @@
 'use strict';
 
-(function ($) {
+(function($) {
 
-  $.fn.autocompleteselect = function (options) {
-    return this.each(function () {
+  $.fn.autocompleteselect = function(options) {
+    return this.each(function() {
       var id = this.id,
           $this = $(this),
           $text = $('#' + id + '_text'),
@@ -23,15 +23,15 @@
       }
 
       function addKiller(repr, pk) {
-        var killer_id = 'kill_' + pk + id,
-            killButton = '<span class="ui-icon ui-icon-trash" id="' + killer_id + '">X</span> ';
+        var killId = 'kill_' + pk + id,
+            killButton = '<span class="ui-icon ui-icon-trash" id="' + killId + '">X</span> ';
         if (repr) {
           $deck.empty();
           $deck.append('<div>' + killButton + repr + '</div>');
         } else {
-          $('#' + id+'_on_deck > div').prepend(killButton);
+          $('#' + id + '_on_deck > div').prepend(killButton);
         }
-        $('#' + killer_id).click(function () {
+        $('#' + killId).click(function() {
           kill();
           $deck.trigger('killed', [pk]);
         });
@@ -45,28 +45,36 @@
       options.select = receiveResult;
       $text.autocomplete(options);
 
-      if (options.initial) {
-        addKiller(options.initial[0], options.initial[1]);
+      function reset() {
+        if (options.initial) {
+          addKiller(options.initial[0], options.initial[1]);
+          $this.val(options.initial[1]);
+        } else {
+          kill();
+        }
       }
 
-      $this.bind('didAddPopup', function (event, pk, repr) {
+      reset();
+      $this.closest('form').on('reset', reset);
+
+      $this.bind('didAddPopup', function(event, pk, repr) {
         receiveResult(null, {item: {pk: pk, repr: repr}});
       });
     });
   };
 
-  $.fn.autocompleteselectmultiple = function (options) {
-    return this.each(function () {
+  $.fn.autocompleteselectmultiple = function(options) {
+    return this.each(function() {
       var id = this.id,
           $this = $(this),
-          $text = $('#' + id+'_text'),
-          $deck = $('#' + id+'_on_deck');
+          $text = $('#' + id + '_text'),
+          $deck = $('#' + id + '_on_deck');
 
       function receiveResult(event, ui) {
         var pk = ui.item.pk,
             prev = $this.val();
 
-        if (prev.indexOf('|'+pk+'|') === -1) {
+        if (prev.indexOf('|' + pk + '|') === -1) {
           $this.val((prev ? prev : '|') + pk + '|');
           addKiller(ui.item.repr, pk);
           $text.val('');
@@ -77,11 +85,11 @@
       }
 
       function addKiller(repr, pk) {
-        var killer_id = 'kill_' + pk + id,
-            killButton = '<span class="ui-icon ui-icon-trash" id="' + killer_id + '">X</span> ';
+        var killId = 'kill_' + pk + id,
+            killButton = '<span class="ui-icon ui-icon-trash" id="' + killId + '">X</span> ';
         $deck.append('<div id="' + id + '_on_deck_' + pk + '">' + killButton + repr + ' </div>');
 
-        $('#' + killer_id).click(function () {
+        $('#' + killId).click(function() {
           kill(pk);
           $deck.trigger('killed', [pk]);
         });
@@ -89,20 +97,29 @@
 
       function kill(pk) {
         $this.val($this.val().replace('|' + pk + '|', '|'));
-        $('#' + id+'_on_deck_'+pk).fadeOut().remove();
+        $('#' + id + '_on_deck_' + pk).fadeOut().remove();
       }
 
       options.select = receiveResult;
       $text.autocomplete(options);
 
-      if (options.initial) {
-        $.each(options.initial, function (i, its) {
-          addKiller(its[0], its[1]);
-        });
+      function reset() {
+        $deck.empty();
+        var query = '|';
+        if (options.initial) {
+          $.each(options.initial, function(i, its) {
+            addKiller(its[0], its[1]);
+            query += its[1] + '|';
+          });
+        }
+        $this.val(query);
       }
 
-      $this.bind('didAddPopup', function (event, pk, repr) {
-        receiveResult(null, {item: {pk: pk, repr: repr }});
+      reset();
+      $this.closest('form').on('reset', reset);
+
+      $this.bind('didAddPopup', function(event, pk, repr) {
+        receiveResult(null, {item: {pk: pk, repr: repr}});
       });
     });
   };
@@ -172,12 +189,12 @@
   };
 
   // activate any on page
-  $(window).bind('init-autocomplete', function () {
+  $(window).bind('init-autocomplete', function() {
 
-    $('input[data-ajax-select=autocomplete]').each(function (i, inp) {
-      addAutoComplete(inp, function ($inp, opts) {
+    $('input[data-ajax-select=autocomplete]').each(function(i, inp) {
+      addAutoComplete(inp, function($inp, opts) {
         opts.select =
-            function (event, ui) {
+            function(event, ui) {
               $inp.val(ui.item.value).trigger('added', [ui.item.pk, ui.item]);
               return false;
             };
@@ -185,27 +202,28 @@
       });
     });
 
-    $('input[data-ajax-select=autocompleteselect]').each(function (i, inp) {
-      addAutoComplete(inp, function ($inp, opts) {
+    $('input[data-ajax-select=autocompleteselect]').each(function(i, inp) {
+      addAutoComplete(inp, function($inp, opts) {
         $inp.autocompleteselect(opts);
       });
     });
 
-    $('input[data-ajax-select=autocompleteselectmultiple]').each(function (i, inp) {
-      addAutoComplete(inp, function ($inp, opts) {
+    $('input[data-ajax-select=autocompleteselectmultiple]').each(function(i, inp) {
+      addAutoComplete(inp, function($inp, opts) {
         $inp.autocompleteselectmultiple(opts);
       });
     });
 
   });
 
-  $(document).ready(function () {
+  $(document).ready(function() {
     // if dynamically injecting forms onto a page
     // you can trigger them to be ajax-selects-ified:
     $(window).trigger('init-autocomplete');
-    $('.inline-group ul.tools a.add, .inline-group div.add-row a, .inline-group .tabular tr.add-row td a').on('click', function() {
-      $(window).trigger('init-autocomplete');
-    });
+    $('.inline-group ul.tools a.add, .inline-group div.add-row a, .inline-group .tabular tr.add-row td a')
+      .on('click', function() {
+        $(window).trigger('init-autocomplete');
+      });
   });
 
 })(window.jQuery);